Slight improvement to StateVariable
